% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmss-package.R
\docType{package}
\name{brmss-package}
\alias{brmss-package}
\title{brmss: Bayesian Regression Models using Spike-and-Slab Priors}
\description{
Fit high-dimensional Bayesian linear and nonlinear regression models using independent spike-and-slab priors for full Bayesian inference. A general Metropolis-Hastings algorithm is used to sample the indicator variables of variable selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ocbe-uio/brmss}
  \item Report bugs at \url{https://github.com/ocbe-uio/brmss/issues}
}

}
\author{
\strong{Maintainer}: Zhi Zhao \email{zhi.zhao@medisin.uio.no}

}
\keyword{internal}
