% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmss.R
\name{brmss}
\alias{brmss}
\title{Fit Bayesian Regression Models}
\usage{
brmss(
  y,
  x,
  family = "weibull",
  nIter = 500,
  burnin = 200,
  thin = 1,
  tick = 100,
  hyperpar = NULL,
  threads = 1,
  gammaSampler = "mc3",
  initial = NULL,
  arms.list = NULL
)
}
\arguments{
\item{y}{response variable. A vector, matrix or dataframe}

\item{x}{input matrix or dataframe}

\item{family}{a character string representing one of the built-in families}

\item{nIter}{the number of iterations of the chain}

\item{burnin}{number of iterations to discard at the start of the chain}

\item{thin}{thinning MCMC intermediate results to be stored}

\item{tick}{an integer used for printing the iteration index and some updated
parameters every tick-th iteration. Default is 1}

\item{hyperpar}{a list of relevant hyperparameters}

\item{threads}{number of threads used for parallelization. Default is 1}

\item{gammaSampler}{one of \code{c("mc3", "bandit")}}

\item{initial}{a list of initial values for parameters "kappa" and "betas"}

\item{arms.list}{a list of parameters for the ARMS algorithm}
}
\value{
An object of a list including the following components:
\itemize{
\item input - a list of all input parameters by the user
\item output - a list of the all mcmc output estimates:
\itemize{
\item "\code{kappa}" - a vector with MCMC intermediate estimates of the Weibull's shape parameter
\item "\code{betas}" - a matrix with MCMC intermediate estimates of effects on cluster-specific survival
\item "\code{gammas}" - a matrix with MCMC intermediate estimates of inclusion indicators of variables for cluster-specific survival
\item "\code{gamma_acc_rate}" - acceptance rate of the M-H sampling for gammas
\item "\code{loglikelihood}" - a matrix with MCMC intermediate estimates of individuals' likelihoods
\item "\code{tauSq}" - a vector with MCMC intermediate estimates of tauSq
\item "\code{post}" - a list with posterior means of "kappa", "betas", "gammas",
}
\item call - the matched call
}
}
\description{
This is the main function to fit various Bayesian regression models
}
\details{
family TODO:.....
}
\examples{

# simulate data
set.seed(123)
n <- 200 # subjects
p <- 10 # variable selection predictors

dat <- simData(n, p)

# run a Bayesian brmss model: brmss-Ber2
fit <- brmss(dat$y, dat$x, family = "weibull", nIter = 100, burnin = 10)

}
\references{
Eddelbuettel D, Sanderson C (2014). \emph{RcppArmadillo: Accelerating R with high-performance C++ linear algebra}. Computational Statistics and Data Analysis, 71, 1054--1063

Zhao Z (2025+). arXiv.
}
