% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBrier.R
\name{plotBrier}
\alias{plotBrier}
\title{Plot curves of time-dependent Brier score}
\usage{
plotBrier(
  dat,
  datMCMC,
  dat.new = NULL,
  time.star = NULL,
  xlab = "Time",
  ylab = "Brier score",
  ...
)
}
\arguments{
\item{dat}{input data as a list containing outcomes and covariates}

\item{datMCMC}{returned object from the main function \code{brmss()}}

\item{dat.new}{input data for out-sample prediction, with the same format
as \code{dat}}

\item{time.star}{largest time for survival prediction}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{...}{other parameters}
}
\value{
A \code{ggplot2::ggplot} object. See \code{?ggplot2::ggplot} for more
details of the object.
}
\description{
Predict time-dependent Brier scores based on different survival models
}
\examples{

# simulate data
set.seed(123)
n <- 200 # subjects
p <- 10 # variable selection predictors

dat <- simData(n, p, model = "weibull")

# run a Bayesian brmss model
fit <- brmss(dat$y, dat$x, family = "weibull", nIter = 100, burnin = 10)

plotBrier(dat, datMCMC = fit)

}
