% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoeff.R
\name{plotCoeff}
\alias{plotCoeff}
\title{Plot posterior estimates of regression coefficients}
\usage{
plotCoeff(
  dat,
  datMCMC,
  estimator = "beta",
  intercept = FALSE,
  bandwidth = NULL,
  xlim = NULL,
  xlab = NULL,
  label.y = NULL,
  first.coef = NULL,
  y.axis.size = 8,
  ...
)
}
\arguments{
\item{dat}{input data as a list containing outcomes and covariates}

\item{datMCMC}{returned object from the main function \code{brmss()}}

\item{estimator}{print estimators, one of \code{c("beta", "gamma")}}

\item{intercept}{logical value to print intercepts}

\item{bandwidth}{a value of bandwidth used for the ridgeplot}

\item{xlim}{numeric vectors of length 2, giving the x-coordinate range.}

\item{xlab}{a title for the x axis}

\item{label.y}{a title for the y axis}

\item{first.coef}{number of the first variables. Default \code{NULL} for
all variables}

\item{y.axis.size}{text size in pts}

\item{...}{others}
}
\value{
A \code{ggplot2::ggplot} object. See \code{?ggplot2::ggplot} for more
details of the object.
}
\description{
create nice plots for estimated coefficients and 95% credible intervals
}
\examples{

# simulate data
set.seed(123)
n <- 200 # subjects
p <- 10 # variable selection predictors

dat <- simData(n, p)

# run a Bayesian brmss model: brmss-Ber2
fit <- brmss(dat$y, dat$x, family = "weibull", nIter = 100, burnin = 10)

plotCoeff(dat, datMCMC = fit, estimator = "beta")

}
