% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.brmss.R
\name{predict.brmss}
\alias{predict.brmss}
\title{Prediction of survival probability}
\usage{
\method{predict}{brmss}(object, dat, newdata = NULL, type = "survival", times = NULL, ...)
}
\arguments{
\item{object}{the results of a \code{brmss} fit}

\item{dat}{the dataset used in \code{brmss()}}

\item{newdata}{optional new data at which to do predictions. If missing, the
prediction will be based on the training data}

\item{type}{the type of predicted value. Currently it is only valid with
\code{'survival'}}

\item{times}{evaluation time points for survival prediction. Default
\code{NULL} for predicting all time points in the \code{newdata} set}

\item{...}{for future methods}
}
\value{
A matrix object
}
\description{
Compute predicted survival probability for a brmss
}
\examples{

# simulate data
set.seed(123)
n <- 200 # subjects
p <- 10 # variable selection predictors

dat <- simData(n, p, model = "weibull")

# run a Bayesian brmss model
fit <- brmss(dat$y, dat$x, family = "weibull", nIter = 100, burnin = 10)

pred.survival <- predict(fit, dat, newdata = dat, times = c(1, 3, 5))

}
