% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Main function implemented in C++ for the MCMC loop}
\usage{
run_mcmc(
  y,
  X,
  family,
  nIter,
  burnin,
  thin,
  tick,
  gamma_sampler,
  gammaProposal,
  gamma_gibbs,
  var_prior,
  threads,
  n,
  nsamp,
  ninit,
  convex,
  npoint,
  initList,
  rangeList,
  hyperparList
)
}
\arguments{
\item{y}{response variable. A vector, matrix or dataframe}

\item{X}{input matrix or dataframe}

\item{family}{a character string representing one of the built-in families}

\item{nIter}{the number of iterations of the chain}

\item{burnin}{number of iterations to discard at the start of the chain}

\item{thin}{thinning MCMC intermediate results to be stored}

\item{tick}{an integer used for printing the iteration index and some updated
parameters every tick-th iteration. Default is 1}

\item{gamma_sampler}{M-H sampler with "mc" or multi-armed "bandit" proposal for gammas}

\item{gammaProposal}{one of 'c("simple", "posterior")'}

\item{gamma_gibbs}{one of 'c("none", "independent", "gprior")'}

\item{var_prior}{string indicating the prior for the variance of response/error term}

\item{threads}{number of threads used for parallelization. Default is 1}

\item{n}{number of samples to draw}

\item{nsamp}{how many samples to draw for generating each sample; only the last draw will be kept}

\item{ninit}{number of initials as meshgrid values for envelop search}

\item{convex}{adjustment for convexity (non-negative value, default 1.0)}

\item{npoint}{maximum number of envelope points}

\item{initList}{a list of initial values for parameters "kappa", "betas"}

\item{rangeList}{a list of ranges of initial values for parameters "kappa", "betas"}

\item{hyperparList}{a list of relevant hyperparameters}
}
\description{
Main function implemented in C++ for the MCMC loop
}
