% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulate data}
\usage{
simData(n = 200, p = 10, L = 1, kappas = 2, model = "weibull")
}
\arguments{
\item{n}{number of subjects}

\item{p}{number of covariates in each cluster}

\item{L}{number of response variables}

\item{kappas}{value of the Weibull's shape parameter}

\item{model}{one of \code{c("gaussian", "logit", "probit", "poisson",
"beta", "multinomial", "dirichlet", "cox")}}
}
\value{
An object of a list
\itemize{
\item "\code{x}" - an matrix of covariates
\item "\code{y}" - a dataframe including events and times
\item "\code{betas}" - covariate effects
\item "\code{kappas}" - value of the Weibull's shape parameter
}
}
\description{
Simulate survival data based on a Cox model
}
\examples{

# simulate data
set.seed(123)
n <- 200 # subjects
p <- 10 # variable selection predictors
dat <- simData(n, p, model = "weibull")
str(dat)

}
