% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Target density}
\usage{
target(x, mu, kappas)
}
\arguments{
\item{x}{value generated from the proposal distribution}

\item{mu}{mean survival time}

\item{kappas}{Weibull's true shape parameter}
}
\value{
value of the targeted (improper) probability density function
}
\description{
Predefined target density corresponding to the population survival function
of brmss
}
\examples{

time1 <- target(1.2, 0.1, 2)

}
