% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmss.R
\name{brmss}
\alias{brmss}
\title{Fit Bayesian Regression Models}
\usage{
brmss(
  y,
  x,
  family = "weibull",
  nIter = 500,
  burnin = 200,
  thin = 1,
  tick = 100,
  hyperpar = NULL,
  threads = 1,
  gammaSampler = "bandit",
  gammaProposal = "posterior",
  gammaGibbs = "none",
  varPrior = "IG",
  RW.MH = "adaptive",
  initial = NULL,
  arms.list = NULL
)
}
\arguments{
\item{y}{response variable. A vector, matrix or dataframe. For survival 
outcomes, the first column of \code{y} is times and the second column is 
events}

\item{x}{input matrix or dataframe}

\item{family}{a character string representing one of the built-in families}

\item{nIter}{the number of iterations of the chain}

\item{burnin}{number of iterations to discard at the start of the chain}

\item{thin}{thinning MCMC intermediate results to be stored}

\item{tick}{an integer used for printing the iteration index and some updated
parameters every tick-th iteration. Default is 1}

\item{hyperpar}{a list of relevant hyperparameters}

\item{threads}{number of threads used for parallelization. Default is 1}

\item{gammaSampler}{one of \code{c("mc3", "bandit")}}

\item{gammaProposal}{one of \code{c("simple", "posterior")}}

\item{gammaGibbs}{one of \code{c("none", "independent", "gprior", "mrf")}. If 
\code{gammaGibbs = "independent"}, it implements the approach in 
Kuo and Mallick (1998) with gibbs sampling for gammas and with independent 
spike-and-slab priors for betas. If \code{gammaGibbs = "gprior"}, it 
implements the approach in George and McCulloch (1997) with gibbs sampling 
for gammas and with g-prior for betas independent spike-and-slab priors for 
betas.}

\item{varPrior}{string indicating the prior for the variance of 
response/error term. Default is \code{IG}. For continous multivariate 
responses, it can also be \code{IW} for the inverse-Wishart prior (dense 
variance-covariance matrix) or \code{HIW} for the hyper-inverse Wishart 
(sparse precision matrix)}

\item{RW.MH}{string indicating the type of random-walk variance in MH 
sampling for gamma-beta move, one of 
\code{c("fisher", "adaptive", "symmetric")}}

\item{initial}{a list of initial values for parameters "kappa" and "betas"}

\item{arms.list}{a list of parameters for the ARMS algorithm}
}
\value{
An object of a list including the following components:
\itemize{
\item input - a list of all input parameters by the user
\item output - a list of the all mcmc output estimates:
\itemize{
\item "\code{kappa}" - a vector with MCMC intermediate estimates of the Weibull's shape parameter
\item "\code{betas}" - a matrix with MCMC intermediate estimates of effects on cluster-specific survival
\item "\code{gammas}" - a matrix with MCMC intermediate estimates of inclusion indicators of variables for cluster-specific survival
\item "\code{gamma_acc_rate}" - acceptance rate of the M-H sampling for gammas
\item "\code{loglikelihood}" - a matrix with MCMC intermediate estimates of individuals' likelihoods
\item "\code{tauSq}" - a vector with MCMC intermediate estimates of tauSq
\item "\code{sigmaSq}" - a vector with MCMC intermediate estimates of sigmaSq
\item "\code{post}" - a list with posterior means of "kappa", "betas", "gammas" and sigmaSq"
}
\item call - the matched call
}
}
\description{
This is the main function to fit various Bayesian regression models
}
\examples{

# simulate data
set.seed(123)
n <- 200 # subjects
p <- 10 # variable selection predictors

dat <- simData(n, p, model = "weibull")

# run a Bayesian brmss model
fit <- brmss(dat$y, dat$x, family = "weibull", nIter = 10, burnin = 1)

}
\references{
Eddelbuettel D, Sanderson C (2014). \emph{RcppArmadillo: Accelerating R with high-performance C++ linear algebra}. Computational Statistics and Data Analysis, 71, 1054--1063

Zhao Z (2025+). arXiv.
}
