% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMC.R
\name{plotMCMC}
\alias{plotMCMC}
\title{MCMC trace-plots}
\usage{
plotMCMC(dat, datMCMC, estimator = "tau")
}
\arguments{
\item{dat}{input data as a list containing outcomes and covariates}

\item{datMCMC}{returned object from the main function \code{brmss()}}

\item{estimator}{print estimators, one of
\code{c("beta", "gamma", "tau", "sigma", "kappa")}}
}
\value{
A \code{ggplot2::ggplot} object. See \code{?ggplot2::ggplot} for more
details of the object.
}
\description{
Trace-plots of regression coefficients over MCMC iterations
}
\examples{

# simulate data
set.seed(123)
n <- 200 # subjects
p <- 10 # variable selection predictors

dat <- simData(n, p, model = "weibull")

# run a Bayesian brmss model
fit <- brmss(dat$y, dat$x, family = "weibull", nIter = 100, burnin = 10)

plotMCMC(dat, datMCMC = fit, estimator = "tau")

}
